requirements.txt
numpy
pandas
accelerate
evaluate
datasets
ipykernel
ipython
loguru
peft
torch
trl
transformers
tqdm
icecream
python-dotenv
tensorboard
bitsandbytes
magic-wormhole
packaging
hydra-core
lean-dojo
vllm
editdistance
# flash-attn instructions: 
# - apparently you need to install everything and then flash-attn
# - make sure to have flash-attn commented on the first pass
# - if in a brand new environment, you may need to install cuda-nvcc to make sure flash-attn installs correctly
# flash-attn
